+++ src/satellite/src/storage/mod.rs
@@ -3,7 +3,7 @@ pub mod constants;
 pub mod custom_domains;
 pub mod http;
 pub mod impls;
-mod runtime;
+pub mod runtime;
 mod state;
 pub mod store;
 pub mod types;
